requirements.txt
SQLAlchemy==2.0.34
pymysql==1.1.2
openpyxl==3.1.5
flask==3.0.3
pandas==2.2.3
python-dotenv==1.0.1